CrunchyRest/requirements.txt
djangorestframework==3.14.0
djongo==1.3.6
pymongo[srv]==3.12.1
pika==1.3.2
python-decouple==3.8
django-cors-headers==4.1.0
neo4j==5.10.0
tqdm==4.65.0
pycountry==24.6.1
babel==2.16.0
requests==2.31.0
regex
django-pytest
# Entity resolution and cross-discovery dependencies
tldextract==5.1.2
thefuzz[speedup]==0.22.1
ddgs>=7.0.0